2005-10-17  Christophe Fergeau  <teuf@gnome.org>
	* src/db-image-parser.c: (get_pixel_data),
	(ipod_image_new_from_mhni):
	* src/itdb.h:
	* src/itdb_itunesdb.c: (get_mhod): Changed type of the various fields
	in the ItdbImage structures so that they match what they are in the
	database on the iPod.


git-svn-id: https://gtkpod.svn.sourceforge.net/svnroot/gtkpod/libgpod/trunk@1124 f01d2545-417e-4e96-918e-98f8d0dbbcb6

+++ src/itdb.h
@@ -320,10 +320,10 @@ enum ItdbImageType {
 struct _Itdb_Image {
 	enum ItdbImageType type;
   	char *filename;
-	off_t offset;
-	size_t size;
-	unsigned int width;
-	unsigned int height;
+	guint32 offset;
+	guint32 size;
+	gint16 width;
+	gint16 height;
 };
 
 typedef struct _Itdb_Image Itdb_Image;
